requirements.txt
# requirements.txt
django>=4.2
qrcode
pillow